Convert all expiration values to long to prevent undesirable calculation results

+++ Server/src/main/java/org/gusdb/oauth2/server/TokenExpirerThread.java
@@ -24,9 +24,9 @@ public class TokenExpirerThread implements Runnable {
     _executor.shutdown();
   }
 
-  private int _tokenExpirationSecs;
+  private long _tokenExpirationSecs;
 
-  public TokenExpirerThread(int tokenExpirationSecs) {
+  public TokenExpirerThread(long tokenExpirationSecs) {
     _tokenExpirationSecs = tokenExpirationSecs;
   }
